+++ modules/DataLayer/Storage/base64.h
@@ -3,7 +3,7 @@
 #include <string>
 
 std::string base64_encode( char const* );
-std::string base64_encode( char const* , unsigned int len);
+std::string base64_encode( char const* , size_t len);
 std::string base64_decode(std::string const& s);
 inline std::string base64_decode( const char *data) {
   return data ? base64_decode( std::string( data)) : "(null)";
